requirements.txt
django-crispy-forms
Django
djangorestframework
flake8
ipdb